+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
 Fixed
 -----
 - Date selection for word statistics was broken
+- Incorrect datatype for column in user_events table
 
 ----------
 `0.7.0`_ - 2023-01-14